@@ -2,48 +2,48 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": ">= 8.0",
+        "php": ">= 8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "antishov/doctrine-extensions-bundle": "^1.4",
         "api-platform/core": "^2.6",
-        "composer/package-versions-deprecated": "1.11.99.1",
+        "composer/package-versions-deprecated": ">=1.11.99.5",
         "doctrine/annotations": "^1.0",
         "doctrine/doctrine-bundle": "^2.2",
         "doctrine/doctrine-migrations-bundle": "^3.0",
         "doctrine/orm": "^2.8",
-        "easycorp/easyadmin-bundle": "^2.3",
-        "friendsofsymfony/jsrouting-bundle": "^2.5",
-        "knplabs/knp-markdown-bundle": "^1.8",
+        "doctrine/persistence": "^2.2",
+        "friendsofsymfony/jsrouting-bundle": "^3.1.1",
         "nelmio/cors-bundle": "^2.1",
         "phpdocumentor/reflection-docblock": "^5.2",
-        "sensio/framework-extra-bundle": "^6.2",
-        "survos/base-bundle": "^1.0@dev",
-        "symfony/asset": "^5.0",
-        "symfony/console": "^5.0",
-        "symfony/dotenv": "^5.0",
-        "symfony/expression-language": "^5.0",
-        "symfony/flex": "^1.3.1",
-        "symfony/form": "^5.0",
-        "symfony/framework-bundle": "^5.0",
-        "symfony/http-client": "^5.0",
-        "symfony/intl": "^5.0",
-        "symfony/mailer": "^5.0",
+        "sensio/framework-extra-bundle": "^6.0",
+        "stof/doctrine-extensions-bundle": "^1.4",
+        "survos/base-bundle": "dev-master",
+        "survos/workflow-extension-bundle": "*@dev",
+        "symfony/asset": "^6.0",
+        "symfony/console": "^6.0",
+        "symfony/dotenv": "^6.0",
+        "symfony/expression-language": "^6.0",
+        "symfony/flex": "^2.0",
+        "symfony/form": "^6.0",
+        "symfony/framework-bundle": "^6.0",
+        "symfony/http-client": "^6.0",
+        "symfony/intl": "^6.0",
+        "symfony/mailer": "^6.0",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/notifier": "^5.0",
-        "symfony/process": "^5.0",
-        "symfony/property-access": "^5.3",
-        "symfony/property-info": "^5.3",
-        "symfony/proxy-manager-bridge": "^5.3",
-        "symfony/security-bundle": "^5.0",
-        "symfony/serializer": "^5.3",
-        "symfony/string": "^5.0",
-        "symfony/translation": "^5.0",
-        "symfony/twig-bundle": "^5.3",
-        "symfony/uid": "^5.3",
-        "symfony/validator": "^5.0",
-        "symfony/web-link": "^5.0",
-        "symfony/yaml": "^5.0",
+        "symfony/notifier": "^6.0",
+        "symfony/process": "^6.0",
+        "symfony/property-access": "^6.0",
+        "symfony/property-info": "^6.0",
+        "symfony/proxy-manager-bridge": "^6.0",
+        "symfony/security-bundle": "^6.0",
+        "symfony/serializer": "^6.0",
+        "symfony/string": "^6.0",
+        "symfony/translation": "^6.0",
+        "symfony/twig-bundle": "^6.0",
+        "symfony/uid": "^6.0",
+        "symfony/validator": "^6.0",
+        "symfony/web-link": "^6.0",
+        "symfony/yaml": "^6.0",
         "twig/cssinliner-extra": "^3.0",
         "twig/extra-bundle": "^2.12|^3.0",
         "twig/inky-extra": "^3.0",
@@ -54,19 +54,24 @@
     "require-dev": {
         "doctrine/doctrine-fixtures-bundle": "^3.3",
         "phpunit/phpunit": "^9.5",
-        "symfony/browser-kit": "^5.3",
-        "symfony/css-selector": "^5.3",
-        "symfony/debug-bundle": "^5.3",
+        "rector/rector": "^0.12.21",
+        "survos/maker-bundle": "dev-main",
+        "symfony/browser-kit": "^6.0",
+        "symfony/css-selector": "^6.0",
+        "symfony/debug-bundle": "^6.0",
         "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "^5.3",
-        "symfony/stopwatch": "^5.3",
-        "symfony/web-profiler-bundle": "^5.3"
+        "symfony/phpunit-bridge": "^6.0",
+        "symfony/stopwatch": "^6.0",
+        "symfony/web-profiler-bundle": "^6.0"
     },
     "config": {
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "symfony/flex": true
+        }
     },
     "autoload": {
         "psr-4": {
