@@ -12,48 +12,39 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-/**
- * @Route("/building")
- */
+#[Route(path: '/building')]
 class BuildingController extends AbstractController
 {
-    /**
-     * @Route("/", name="building_index", methods={"GET"})
-     */
-    public function index(BuildingRepository $buildingRepository): Response
+    public function __construct(private EntityManagerInterface $entityManager) {
+
+    }
+    #[Route(path: '/', name: 'building_index', methods: ['GET'])]
+    public function index(BuildingRepository $buildingRepository) : Response
     {
         return $this->render('building/index.html.twig', [
             'buildings' => $buildingRepository->findAll(),
         ]);
     }
-
-    /**
-     * @Route("/new", name="building_new", methods={"GET","POST"})
-     */
-    public function new(Request $request): Response
+    #[Route(path: '/new', name: 'building_new', methods: ['GET', 'POST'])]
+    public function new(Request $request) : Response
     {
         $building = new Building();
         $form = $this->createForm(BuildingType::class, $building);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager = $this->entityManager;
             $entityManager->persist($building);
             $entityManager->flush();
 
             return $this->redirectToRoute('building_index');
         }
-
         return $this->render('building/new.html.twig', [
             'building' => $building,
             'form' => $form->createView(),
         ]);
     }
-
-    /**
-     * @Route("/{id}", name="building_show", methods={"GET"})
-     */
-    public function show(Building $building, EntityManagerInterface $entityManager): Response
+    #[Route(path: '/{id}', name: 'building_show', methods: ['GET'])]
+    public function show(Building $building, EntityManagerInterface $entityManager) : Response
     {
         $repo = $entityManager->getRepository(Location::class);
         return $this->render('building/show.html.twig', [
@@ -62,38 +53,29 @@ class BuildingController extends AbstractController
             'building' => $building,
         ]);
     }
-
-    /**
-     * @Route("/{id}/edit", name="building_edit", methods={"GET","POST"})
-     */
-    public function edit(Request $request, Building $building): Response
+    #[Route(path: '/{id}/edit', name: 'building_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Building $building) : Response
     {
         $form = $this->createForm(BuildingType::class, $building);
         $form->handleRequest($request);
-
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $this->entityManager->flush();
 
             return $this->redirectToRoute('building_index');
         }
-
         return $this->render('building/edit.html.twig', [
             'building' => $building,
             'form' => $form->createView(),
         ]);
     }
-
-    /**
-     * @Route("/{id}", name="building_delete", methods={"DELETE"})
-     */
-    public function delete(Request $request, Building $building): Response
+    #[Route(path: '/{id}', name: 'building_delete', methods: ['DELETE'])]
+    public function delete(Request $request, Building $building) : Response
     {
         if ($this->isCsrfTokenValid('delete'.$building->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager = $this->entityManager;
             $entityManager->remove($building);
             $entityManager->flush();
         }
-
         return $this->redirectToRoute('building_index');
     }
 }
