@@ -11,167 +11,132 @@ use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Mapping\Annotation as Gedmo;
 
 /**
- * @ORM\Entity(repositoryClass=TopicRepository::class)
  * @Gedmo\Tree(type="nested")
  */
-class Topic
+#[ORM\Entity(repositoryClass: TopicRepository::class)]
+class Topic implements \Stringable
 {
-
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="string", length=10)
-     */
+    #[ORM\Id]
+    #[ORM\Column(type: 'string', length: 10)]
     private $code;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
+    #[ORM\Column(type: 'string', length: 255)]
     private $name;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
+    #[ORM\Column(type: 'text')]
     private $description;
-
     public function getCode(): ?string
     {
         return $this->code;
     }
-
     public function setCode(string $code): self
     {
         $this->code = $code;
 
         return $this;
     }
-
     public function getName(): ?string
     {
         return $this->name;
     }
-
     public function setName(string $name): self
     {
         $this->name = $name;
 
         return $this;
     }
-
     public function getDescription(): ?string
     {
         return $this->description;
     }
-
     public function setDescription(string $description): self
     {
         $this->description = $description;
 
         return $this;
     }
-
     /**
      * @Gedmo\TreeLeft
-     * @ORM\Column(type="integer")
      */
+    #[ORM\Column(type: 'integer')]
     private $lft;
-
     /**
      * @Gedmo\TreeLevel
-     * @ORM\Column(type="integer")
      */
+    #[ORM\Column(type: 'integer')]
     private $lvl;
-
     /**
      * @Gedmo\TreeRight
-     * @ORM\Column(type="integer")
      */
+    #[ORM\Column(type: 'integer')]
     private $rgt;
-
     /**
      * @Gedmo\TreeRoot
-     * @ORM\ManyToOne(targetEntity="Topic")
-     * @ORM\JoinColumn(referencedColumnName="code", onDelete="CASCADE")
      */
+    #[ORM\ManyToOne(targetEntity: 'Topic')]
+    #[ORM\JoinColumn(referencedColumnName: 'code', onDelete: 'CASCADE')]
     private $root;
-
     /**
      * @Gedmo\TreeParent
-     * @ORM\ManyToOne(targetEntity="Topic", inversedBy="children")
-     * @ORM\JoinColumn(referencedColumnName="code", onDelete="CASCADE")
      */
+    #[ORM\ManyToOne(targetEntity: 'Topic', inversedBy: 'children')]
+    #[ORM\JoinColumn(referencedColumnName: 'code', onDelete: 'CASCADE')]
     private $parent;
-
-    /**
-     * @ORM\OneToMany(targetEntity="Topic", mappedBy="parent")
-     * @ORM\OrderBy({"lft" = "ASC"})
-     */
+    #[ORM\OneToMany(targetEntity: 'Topic', mappedBy: 'parent')]
+    #[ORM\OrderBy(['lft' => 'ASC'])]
     private $children;
-
     public function __construct()
     {
         $this->children = new ArrayCollection();
     }
-
     public function getLft(): ?int
     {
         return $this->lft;
     }
-
     public function setLft(int $lft): self
     {
         $this->lft = $lft;
 
         return $this;
     }
-
     public function getLvl(): ?int
     {
         return $this->lvl;
     }
-
     public function setLvl(int $lvl): self
     {
         $this->lvl = $lvl;
 
         return $this;
     }
-
     public function getRgt(): ?int
     {
         return $this->rgt;
     }
-
     public function setRgt(int $rgt): self
     {
         $this->rgt = $rgt;
 
         return $this;
     }
-
     public function getRoot(): ?Topic
     {
         return $this->root;
     }
-
     public function setRoot(?Topic $root): self
     {
         $this->root = $root;
 
         return $this;
     }
-
     public function getParent(): ?Topic
     {
         return $this->parent;
     }
-
     public function setParent(?Topic $parent): self
     {
         $this->parent = $parent;
 
         return $this;
     }
-
     /**
      * @return Collection|Topic[]
      */
@@ -179,7 +144,6 @@ class Topic
     {
         return $this->children;
     }
-
     public function addChild(Topic $child): self
     {
         if (!$this->children->contains($child)) {
@@ -189,7 +153,6 @@ class Topic
 
         return $this;
     }
-
     public function removeChild(Topic $child): self
     {
         if ($this->children->removeElement($child)) {
@@ -201,15 +164,12 @@ class Topic
 
         return $this;
     }
-
     public function getData()
     {
         return json_encode($this);
     }
-
     public function __toString(): string
     {
         return sprintf("%s %s", $this->getName(), $this->getCode());
     }
-
 }
