@@ -7,62 +7,43 @@ use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Security\Core\User\UserInterface;
 
-/**
- * @ORM\Table(name="users")
- * @ORM\Entity(repositoryClass="App\Repository\UserRepository")
- */
-class User implements UserInterface
+#[ORM\Table(name: 'users')]
+#[ORM\Entity(repositoryClass: 'App\Repository\UserRepository')]
+class User implements UserInterface, \Stringable
 {
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: 'integer')]
     private $id;
-
-    /**
-     * @ORM\Column(type="string", length=180, unique=true)
-     */
+    #[ORM\Column(type: 'string', length: 180, unique: true)]
     private $email;
-
-    /**
-     * @ORM\Column(type="json")
-     */
+    #[ORM\Column(type: 'json')]
     private $roles = [];
-
     /**
      * @var string The hashed password
-     * @ORM\Column(type="string")
      */
+    #[ORM\Column(type: 'string')]
     private $password;
-
-    /**
-     * @ORM\OneToMany(targetEntity="App\Entity\Building", mappedBy="user", orphanRemoval=true, cascade={"persist"})
-     */
+    #[ORM\OneToMany(targetEntity: 'App\Entity\Building', mappedBy: 'user', orphanRemoval: true, cascade: ['persist'])]
     private $buildings;
-
     public function __construct()
     {
         $this->buildings = new ArrayCollection();
     }
-
     public function getId(): ?int
     {
         return $this->id;
     }
-
-    public function getEmail(): ?string
+    public function getEmail(): string
     {
         return $this->email;
     }
-
     public function setEmail(string $email): self
     {
         $this->email = $email;
 
         return $this;
     }
-
     /**
      * A visual identifier that represents this user.
      *
@@ -72,7 +53,6 @@ class User implements UserInterface
     {
         return (string) $this->email;
     }
-
     /**
      * @see UserInterface
      */
@@ -84,14 +64,12 @@ class User implements UserInterface
 
         return array_unique($roles);
     }
-
     public function setRoles(array $roles): self
     {
         $this->roles = $roles;
 
         return $this;
     }
-
     /**
      * @see UserInterface
      */
@@ -99,14 +77,12 @@ class User implements UserInterface
     {
         return (string) $this->password;
     }
-
     public function setPassword(string $password): self
     {
         $this->password = $password;
 
         return $this;
     }
-
     /**
      * @see UserInterface
      */
@@ -114,7 +90,6 @@ class User implements UserInterface
     {
         // not needed when using the "bcrypt" algorithm in security.yaml
     }
-
     /**
      * @see UserInterface
      */
@@ -123,7 +98,6 @@ class User implements UserInterface
         // If you store any temporary, sensitive data on the user, clear it here
         // $this->plainPassword = null;
     }
-
     /**
      * @return Collection|Building[]
      */
@@ -131,7 +105,6 @@ class User implements UserInterface
     {
         return $this->buildings;
     }
-
     public function addBuilding(Building $building): self
     {
         if (!$this->buildings->contains($building)) {
@@ -141,7 +114,6 @@ class User implements UserInterface
 
         return $this;
     }
-
     public function removeBuilding(Building $building): self
     {
         if ($this->buildings->contains($building)) {
@@ -154,9 +126,13 @@ class User implements UserInterface
 
         return $this;
     }
-
-    public function __toString()
+    public function __toString(): string
     {
         return $this->getEmail();
+    }
+    public function getUserIdentifier(): string
+    {
+        return $this->getEmail();
+        // TODO: Implement getUserIdentifier() method.
     }
 }
