@@ -1,6 +1,6 @@
 const $ = require('jquery');
 require('jstree');
-require('jquery-confirm');
+// require('jquery-confirm');
 const routes = require('../../public/js/fos_js_routes.json');
 import Routing from '../../vendor/friendsofsymfony/jsrouting-bundle/Resources/public/js/router.min.js';
 Routing.setRoutingData(routes);
