@@ -21,14 +21,12 @@ Encore
      * and one CSS file (e.g. app.css) if your JavaScript imports CSS.
      */
     .addEntry('app', './assets/js/app.js')
-    .addEntry('tree-api', './assets/js/tree-api.js')
-    .addEntry('building_show', './assets/js/building.js')
-    .addEntry('directory', './assets/js/directory.js')
-//    .addEntry('basic-ajax', './assets/js/basic-ajax.js')
     .addEntry('basic-html', './assets/js/basic-html.js')
-//    .addEntry('LocationManager', './assets/js/LocationManagerApp.js')
-    //.addEntry('page1', './assets/js/page1.js')
-    //.addEntry('page2', './assets/js/page2.js')
+    .addEntry('directory', './assets/js/directory.js')
+    .addEntry('building_show', './assets/js/building.js')
+
+    // enables the Symfony UX Stimulus bridge (used in assets/bootstrap.js)
+    .enableStimulusBridge('./assets/controllers.json')
 
     // When enabled, Webpack "splits" your files into smaller pieces for greater optimization.
     .splitEntryChunks()
@@ -77,11 +75,4 @@ Encore
     //.autoProvidejQuery()
 ;
 
-Encore.autoProvideVariables({
-    '$': 'jquery',
-    'jQuery': 'jquery',
-});
-
-const config = Encore.getWebpackConfig();
-config.resolve.symlinks = false;
-module.exports = config;
+module.exports = Encore.getWebpackConfig();
